+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## WIP
+
+### Removed
+
+- Registry class because of it over-engineering propose.   
+
 ## [0.1.0-alpha2] - 2026-04-05
 
 ### Changed